fixing

+++ backend/src/controllers/auth/logout.ts
@@ -1,19 +1,24 @@
 import { sessionModal } from "db/schema/session.js";
 import { Request, Response } from "express";
+import mongoose from "mongoose";
 import { clearAcCookie, clearRefCookie } from "service/session/session_cookies.js";
 
 export const Logout = async (req: Request, res: Response) => {
 
   const { sessionId } = req.session;
+  const { _id } = req.user
 
-  if (!sessionId) {
+  if (!sessionId || !_id) {
     return res.status(401).json({
       message: "Invalid request"
     })
   }
 
-  const session = await sessionModal.findOneAndDelete({ sessionId });
+  console.log(`_id => ${_id}`)
+  console.log(`sessionId => ${sessionId}`)
 
+  const session = await sessionModal.findOneAndDelete({ sessionId: sessionId, user_id: _id });
+  console.log(session)
   if (!session) {
     return res.status(500).json({
       message: "error logging out"
